[Fix] Chatting 13 이하 버전 예외 처리

+++ app/src/main/java/com/dreamteam/sharedream/home/alarm/AlarmAdapter2.kt
@@ -1,6 +1,5 @@
 package com.dreamteam.sharedream.home.alarm
 
-import AlarmPost
 import android.annotation.SuppressLint
 import android.content.Context
 import android.util.Log
@@ -9,32 +8,30 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import coil.load
 import com.dreamteam.sharedream.databinding.MyalarmBinding
-import com.dreamteam.sharedream.home.Search.SeachAdapter
+import com.dreamteam.sharedream.model.AlarmChatData
 import com.google.firebase.Timestamp
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
-import com.google.firebase.storage.ktx.storage
 import java.text.SimpleDateFormat
 import java.util.Date
 
-
-class AlarmPostAdapter(
+class AlarmAdapter2(
     private val context: Context,
-    private val itemClickListener: AlarmPostAdapter.OnItemClickListener
+//    private val itemClickListener: AlarmAdapter2.OnItemClickListener
 ) :
-    RecyclerView.Adapter<AlarmPostAdapter.AlarmHolder>() {
-    private var alarmItem: List<AlarmPost?> = emptyList()
-    interface OnItemClickListener {
-        fun onItemClick(post: AlarmPost)
-    }
-    fun setData(data: List<AlarmPost?>) {
-        alarmItem = data
+    RecyclerView.Adapter<AlarmAdapter2.AlarmHolder2>() {
+    private var chatItem: List<AlarmChatData?> = emptyList()
+//
+//        interface OnItemClickListener {
+//        fun onItemClick(post: AlarmChatData)
+//    }
+    fun setData(data: List<AlarmChatData?>) {
+        chatItem = data
         notifyDataSetChanged()
         Log.d("nyh", "setData: $data")
     }
 
-    @Suppress("DEPRECATION")
-    inner class AlarmHolder(val binding: MyalarmBinding) :
+    inner class AlarmHolder2(val binding: MyalarmBinding) :
         RecyclerView.ViewHolder(binding.root) {
 
         val alarmImage = binding.imageTrade
@@ -42,11 +39,11 @@ class AlarmPostAdapter(
         val alarmNickname = binding.notinickname
         val alarmTime = binding.txtTime
 
-        fun bind(post: AlarmPost, timestamp: Timestamp) {
+        fun bind(chatPost: AlarmChatData, timestamp: Timestamp) {
 
             val date: Date = timestamp.toDate()
             val dateFormat = SimpleDateFormat("yyyy-MM-dd")
-
+//
 //            postDate.text = dateFormat.format(date)
 
             // 2. 현재 날짜, 시간 기준으로 만들기 ( 1시간 전, 2일 전, 1달 전)
@@ -73,38 +70,40 @@ class AlarmPostAdapter(
                     else -> "${dateFormat.format(date)}"
                 }
             alarmTime.text = result
-            alarmTitle.text = post.title
-            alarmNickname.text = post.nickname
+            alarmTitle.text = "채팅신청이 도착했습니다."
+            alarmNickname.text = chatPost.name
 
-            if (post.imgs.isNotEmpty()) {
-                val imageUrl = post.imgs.first()
+            if (chatPost.profileImageUrl?.isNotEmpty() == true) {
+                val imageUrl = chatPost.profileImageUrl.first()
                 alarmImage.load(imageUrl)
             }
         }
 
         @SuppressLint("NotifyDataSetChanged")
         private fun deleteItem() {
-            val post = alarmItem[position]
+            val post = chatItem[position]
             if (post != null) {
                 val myDocuId = post.myDocuId
 
-                if (alarmItem is MutableList) {
-                    (alarmItem as MutableList).removeAt(position)
+                if (chatItem is MutableList) {
+                    (chatItem as MutableList).removeAt(position)
                 }
                 notifyItemRemoved(position)
 
                 val db = Firebase.firestore
-                db.collection("notifyList")
-                    .document(myDocuId)
-                    .delete()
-                    .addOnSuccessListener {
-                        notifyDataSetChanged()
-                        Log.d("nyh", "deleteItem: 삭제완료")
-                        Log.d("nyh", "deleteItem: $myDocuId")
-                    }
-                    .addOnFailureListener { e ->
-                        Log.e("nyh", "deleteItem: $e")
-                    }
+                if (myDocuId != null) {
+                    db.collection("notifyChatList")
+                        .document(myDocuId)
+                        .delete()
+                        .addOnSuccessListener {
+                            notifyDataSetChanged()
+                            Log.d("nyh", "deleteItem: 삭제완료")
+                            Log.d("nyh", "deleteItem: $myDocuId")
+                        }
+                        .addOnFailureListener { e ->
+                            Log.e("nyh", "deleteItem: $e")
+                        }
+                }
             }
         }
 
@@ -115,27 +114,29 @@ class AlarmPostAdapter(
         }
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): AlarmHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): AlarmHolder2 {
         val inflater = LayoutInflater.from(parent.context)
         val binding = MyalarmBinding.inflate(inflater, parent, false)
-        return AlarmHolder(binding)
+        return AlarmHolder2(binding)
     }
 
     override fun getItemCount(): Int {
-        return alarmItem.size
+        return chatItem.size
     }
 
-    override fun onBindViewHolder(holder: AlarmHolder, position: Int) {
-        val post = alarmItem?.getOrNull(position)
-        holder.itemView.setOnClickListener {
-            if (post != null) {
-                itemClickListener.onItemClick(post)
-            }
-        }
+    override fun onBindViewHolder(holder: AlarmAdapter2.AlarmHolder2, position: Int) {
+        val post = chatItem?.getOrNull(position)
+//        holder.itemView.setOnClickListener {
+//            if (post != null) {
+//                itemClickListener.onItemClick(post)
+//            }
+//        }
 
-        if (post != null) {
-            holder.bind(post, post.timestamp)
-            Log.d("nyh", "onBindViewHolder: $alarmItem")
+        if (post != null && chatItem != null) {
+            if (post != null) {
+                holder.bind(post, post.timestamp)
+            }
+            Log.d("nyh", "onBindViewHolder: $chatItem")
         } else {
             Log.e("nyh", "onBindViewHolder: Data is empty at position $position")
         }
